// modules/testimonials/testimonial.types.ts
import type { MediaImage } from "@/modules/media/media.types";

export interface Testimonial {
  id?: string;
  clientName: string;
  clientPosition?: string;
  projectName?: string;
  serviceId?: string;
  message: string;
  rating: number;
  photo?: MediaImage | null;
  isPublished: boolean;
  order: number;
  createdAt?: string | Date;
  updatedAt?: string | Date;
}